Remove importing Extras from the example
Importing Extras from plugins/ folder can have consequences on the
configuration of the plugin that's enabled in the Extras. lazy.nvim
loads plugins alphabetically from plugins/ folder which can mean
that some plugins will have configuration based on the order in
which they are loaded. When configuration options are getting merged
this has no effect. But when configuration option is supposed to
get completely replaced this can be a problem. For more details see
LazyVim/LazyVim/issues/2567.

Extras documentation recommends using :LazyExtras command or importing
extras in config/lazy.lua. Example should be consistent with that
recommendation.