Deperecated syntax in bootstrapping of LazyVim starter
In the current state of the lazy.nvim repository, certain updates have been made. However, there remains an outdated reference to 'vim.loop' for the bootstrapping process of 'lazyvim' in this start repo for LazyVim, despite 'vim.loop' being deprecated. To rectify this, I suggest a minor alteration by replacing it with the following code snippet:

```lua
if not (vim.uv or vim.loop).fs_stat(lazypath) then
    -- bootstrap
do
```